requirements.txt
streamlit>=1.39.0
huggingface_hub
pandas
python-dotenv